customer_support_service/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14,<3.15
psycopg2-binary>=2.9,<3.0
django-cors-headers>=4.3,<4.4
pika>=1.3,<2.0
python-dotenv>=1.0,<2.0
google-generativeai>=0.3,<1.0
drf-spectacular>=0.27,<1.0
